pyproject.toml
[tool.poetry]
name = "ff-server"
version = "0.1.0"
description = ""
authors = ["nnnLik <rslmakhmudov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
uvicorn = "0.27.1"
redis = "5.0.2"
pydantic = "2.6.3"
pydantic-settings = "2.2.1"
fastapi = "0.110.0"
websockets = "^12.0"
logtail-python = "^0.2.10"
ujson = "^5.9.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
